+++ package.json
@@ -2,7 +2,7 @@
   "name": "lb_radio",
   "version": "0.0.0",
   "description": "Utility to submit listens to ListenBrainz done with a dumb radio receiver",
-  "main": "server/server.js",
+  "main": "server/index.js",
   "private": true,
   "type": "module",
   "scripts": {
@@ -11,8 +11,7 @@
     "build:dev": "webpack --mode=development",
     "build:prod": "webpack --mode=production --node-env=production",
     "watch": "webpack --watch",
-    "serve": "node server/server.js --port 8080",
-    "serve:watch": "node --watch server/server.js --port 8080"
+    "serve": "node server/index.js --port 8080"
   },
   "keywords": [
     "listenbrainz",
@@ -21,6 +20,7 @@
   "author": "MaxEtMoritz",
   "license": "MIT",
   "dependencies": {
+    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
     "bootstrap": "^5.3.0",
     "bootstrap-icons": "^1.10.5",
     "express": "^4.18.2",
